requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-admin-interface==0.19.0
drf-yasg==1.20.0
django-mssql-backend==2.8.1